Website_and_Services/requirements.txt
# Python 3.13 Compatible Requirements
# Updated versions for compatibility with Python 3.13

# Web Framework
Flask==3.1.0
Flask-CORS==5.0.0
Werkzeug==3.1.3

# HTTP Client
aiohttp>=3.10.0
aiosignal>=1.3.1
async-timeout>=4.0.3

# Type annotations and validation
annotated-types>=0.6.0
pydantic>=2.10.0
pydantic_core>=2.27.0
typing_extensions>=4.12.0

# Async support
anyio>=4.6.0
sniffio>=1.3.1

# HTTP and networking
attrs>=24.2.0
certifi>=2024.8.30
charset-normalizer>=3.4.0
idna>=3.10
requests>=2.32.3
urllib3>=2.2.3

# Flask utilities
blinker>=1.8.2
click>=8.1.7
itsdangerous>=2.2.0
Jinja2>=3.1.4
MarkupSafe>=3.0.2

# JSON utilities
jsonpatch>=1.33
jsonpointer>=3.0.0

# LangChain ecosystem
langchain>=0.3.0
langsmith>=0.2.0

# Logging
loguru>=0.7.2

# Data manipulation
marshmallow>=3.23.0
numpy>=2.1.0
python-dateutil>=2.9.0

# Environment and configuration
python-dotenv>=1.0.1
PyYAML>=6.0.2

# Database
SQLAlchemy>=2.0.35
greenlet>=3.1.1

# Vector database (use newer pinecone package instead of pinecone-client)
pinecone>=5.4.0

# Utilities
packaging>=24.2
regex>=2024.9.11
six>=1.16.0
tenacity>=9.0.0
tiktoken>=0.8.0
tqdm>=4.67.0

# Data processing
dataclasses-json>=0.6.7
mypy-extensions>=1.0.0

# Network utilities
multidict>=6.1.0
yarl>=1.13.0
frozenlist>=1.4.1
dnspython>=2.7.0

# Type checking
typing-inspect>=0.9.0

# Windows-specific (only installs on Windows)
win32-setctime>=1.1.0; sys_platform == "win32"

# Color support
colorama>=0.4.6